processing/requirements.txt
rasterio
shapely
geopandas
pystac-client
odc-stac
sqlalchemy
psycopg2-binary
geoalchemy2
numpy
pandas
requests
